Connect backend to axiosInstance and fix CORS with FastAPI - Rename `simulations.ts` to `simulation.ts` for consistency
