+++ frontend/.env.template
@@ -1,0 +1,8 @@
+# -------------------------------------------------------------------
+# EarthFrame frontend configuration
+# Copy this file to `.env` and adjust values as needed.
+# -------------------------------------------------------------------
+
+# Backend configuration
+# -------------------------------------------------------------------
+VITE_API_BASE_URL=http://127.0.0.1:8000
