+++ frontend/src/api/axios.ts
@@ -1,9 +1,11 @@
-import axios from 'axios';
+import axios from "axios";
+
+export const API_BASE_URL =
+  import.meta.env.VITE_API_BASE_URL || "http://127.0.0.1:8000";
 
 const axiosInstance = axios.create({
-  baseURL: import.meta.env.VITE_REACT_APP_API_URL || 'http://localhost:8000/api',
+  baseURL: API_BASE_URL,
   timeout: 10000,
-  // You can add headers or interceptors here
 });
 
 export default axiosInstance;
